refactor : mvc 패턴을 기준으로 클래스 분리
