+++ src/main/kotlin/baseball/service/GameService.kt
@@ -1,0 +1,16 @@
+package baseball.service
+
+import baseball.domain.Computer
+import baseball.domain.GameLogic
+
+class GameService(private val computer: Computer) {
+    fun startGame(userInput: Int): Boolean {
+        val gameLogic = GameLogic(computer.computerNum, userInput)
+
+        return gameLogic.checkResult()
+    }
+
+    fun restart() {
+        computer.updateNumber()
+    }
+}
